requirements.txt
black==23.3.0
mlflow==2.3.2
numpy==1.24.3
pyarrow==11.0.0
PyYAML
pandas==2.0.1
xgboost==1.7.5
fastapi==0.95.1
pydantic==1.10.7
uvicorn==0.22.0
bentoml==1.0.5
fastparquet==0.8.3
scikit-learn==1.1.2
